backend/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-js-asset==2.0.0
django-mptt==0.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
